Display labeled scales in test
This change is 90% of the way to correctly naming and labeling scales for presentation.

The underlying algorithm correctly determines the right enharmonic notes for a scale based on a given scale series.

F# Major / Gb Major produces incorrect results though (for instance F# Major should have the (slightly confusing) E# so that it has 6 Sharps and doesnt reuse F  when written)